Add test for potential filter and pruning conflict
Add a unit test to demonstrate a potential conflict between filtering
and pruning behavior. The expectation is that the `filter_me` node and
all of its children are not analyzed for duplication.

Please note that if the following step is skipped in `Flay#prune`:

    self.hashes.delete_if { |_,nodes| nodes.size == 1 }

This test (and all others) pass.